+++ Source/Current/Windows API CodePack/Components/Core/Dialogs/TaskDialogs/TaskDialogDefaultButton.cs
@@ -1,0 +1,5 @@
+namespace Microsoft.WindowsAPICodePack.Dialogs;
+
+internal class TaskDialogDefaultButton
+{
+}