+++ Changelog.md
@@ -1,3 +1,16 @@
+01/03/2024
+
+Version 8.0.3
+- Fix the implementation of the property `CheckSelect`
+
+24/02/2024
+
+Version 8.0.2
+- Fix for multi-dotted extensions
+- Remove the border on the `ExplorerBrowser` control 
+- New `SelectionChangedEventArgs` for selected folder and file names
+- Ensure that `Shell` is thread safe
+
 14/02/2024
 
 Version 8.0.1